Enable -Wextra and -Wpedantic warnings and resolve resulting issues

+++ console.c
@@ -240,7 +240,7 @@ consoleread(struct inode *ip, char *dst, int n)
     }
     c = input.buf[input.r++ % INPUT_BUF];
     if(c == C('D')){  // EOF
-      if(n < target){
+      if((uint)n < target){
         // Save ^D for next time, to make sure
         // caller gets a 0-byte result.
         input.r--;
@@ -255,7 +255,7 @@ consoleread(struct inode *ip, char *dst, int n)
   release(&input.lock);
   ilock(ip);
 
-  return target - n;
+  return (int)(target - (uint)n);
 }
 
 int
